Updated AdditionalHandler. Refs #9480

+++ src/utils/handlers/AdditionalHandler.h
@@ -454,7 +454,7 @@ public:
      * @param[in] voltage Voltage of at connection point for the overhead wire
      * @param[in] currentLimit Current limit of the feeder line
      */
-    virtual void buildTractionSubstation(const CommonXMLStructure::SumoBaseObject* sumoBaseObject, const std::string& id, const float voltage, const float currentLimit) = 0;
+    virtual void buildTractionSubstation(const CommonXMLStructure::SumoBaseObject* sumoBaseObject, const std::string& id, const double voltage, const double currentLimit) = 0;
 
     /** @brief build overhead wire clamp
      * @param[in] id Overhead wire segment ID
